Function to read from can done and working

+++ trunk/car_software/g5/node2/test_vars.h
@@ -2,4 +2,4 @@ extern int tIntReg;
 extern int tIntHjul5;
 extern int tIntHjul6;
 
-extern volatile unsigned int canDataTest[8];
+extern volatile unsigned short int canDataTest[8];
